[SMT] Add concat, extract, repeat operations

+++ lib/Dialect/SMT/SMTTypes.cpp
@@ -32,8 +32,8 @@ void SMTDialect::registerTypes() {
 
 LogicalResult
 BitVectorType::verify(function_ref<InFlightDiagnostic()> emitError,
-                      unsigned width) {
-  if (width <= 0)
+                      uint64_t width) {
+  if (width <= 0U)
     return emitError() << "bit-vector must have at least a width of one";
   return success();
 }
